* FSM/PhatMan/FSM_PhatMan.prs: 	* FSM/PhatMan/Makefile.am: 	* FSM/PhatMan/FSM_PhatMan.prs (added): 	* FSM/PhatMan/PhatMan.cpp (added): 	* FSM/PhatMan/Chorus.cpp (deleted): 	  Rename cpp and add presets. 	* Jeskola/Freeverb/Makefile.am: 	* Jeskola/Freeverb/revmodel.cpp: 	* common/dsplib/dsplib.cpp: 	* Jeskola/Freeverb/Makefile.am (added): 	  Fix the linker error and add missing Makefile.am

+++ Jeskola/Freeverb/revmodel.cpp
@@ -105,7 +105,7 @@ bool revmodel::processreplace(float *inputL, float *inputR, float *outputL, floa
 
 //	memcpy(buffer, inputL, numsamples*sizeof(float)*skip);
 
-	int numsamples2=numsamples;
+	unsigned long numsamples2=numsamples;
     float* outputL2=outputL;
 
 	if (skip==2) {
